Split profile creation into multiple files
TODO: The code needs refactor
