+++ renderer/public/locales/en/profile.json
@@ -70,5 +70,9 @@
       "title": "Profile could not be created!",
       "text": "Please try again later."
     }
+  },
+  "cancelProfileCreation": {
+    "title": "Cancel Profile Creation",
+    "text": "Are you sure you want to cancel profile creation? All changes will be lost."
   }
 }
